[CLN] Extract survey functions to files
